task 14

+++ 14-binary_tree_balance.c
@@ -8,8 +8,8 @@
  */
 int binary_tree_balance(const binary_tree_t *tree)
 {
-	int height_left = 0;
-	int height_right = 0;
+	size_t height_left = 0;
+	size_t height_right = 0;
 
 	if (tree == NULL)
 	{
